requirements.txt
django
djangorestframework
django-allauth
environs
pillow
python-dotenv
requests
cryptography
social-auth-app-django
pipenv
django-cors-headers
python-decouple
psycopg2
psycopg2-binary
djangorestframework-simplejwt
google-auth
google-auth-oauthlib
google-auth-httplib2
gunicorn
whitenoise
django-sendgrid-v5
dj-database-url
config
django-cloudinary-storage
django-filter